+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Update Pandoc to 2.16.1
 - Improve sorting behavior of directories on creation and renaming of files
+- Removed custom middle-click paste code for Linux, cf. #2321
 
 # 2.0.2